backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy>=2.0.37
alembic==1.18.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.22
pydantic-settings==2.13.1
cryptography==46.0.5
python-dateutil==2.9.0.post0
pytz==2026.1

# dev / test
pytest==9.0.2
pytest-cov==7.1.0
